Inelegantly fix graphical glitch
When pieces with the same image pattern (pawns in particular, with the
standard board layout) got sorted contiguously in z-order, the pattern
actually SPANNED the different pieces, because it generated a single
SVG <path> element for them, as an optimization.

This is a simple fix, but like the rest of my patches so far, not very
general of a solution, not good to merge upstream.